[C8-1] Prevent premature OutOfMemoryException when G1 GC invocation is suspended by a long-running native call.
Programs that use the G1 GC could experience spurious OOM exceptions even when the Java heap was far from filling up. This happened when a spin loop that waited for long-running native calls gave up after only two rounds. This small patch makes this loop wait as long as it takes. Typically a few more rounds suffice. Worst case, a full GC would eventually occur (thanks to JDK-8137099) and resolve the situation as well. The patch includes a unit test that provokes needing more than two rounds and succeeds only if the patch is in place.