[C8-15] Fixed JDK-8185005: "Improve performance of ThreadMXBean.getThreadInfo(long ids[], int maxDepth)".
This patch improves the performance a JVM-internal function that looks up a Java Thread instance from an OS thread ID. This benefits several ThreadMXBean calls such as getThreadInfo(), getThreadCpuTime(), and getThreadUserTime(). The relative performance improvement increases with the number of threads in the JVM, as linear search is replaced by a hash table lookup.