Add support for building rpm package on aarch64
redline-1.2.7 is not able to recognize aarch64 as a valid architecture,
causing the rpm gradle task fail silently on aarch64. The issue is fixed
in redline-1.2.8 but the package has not been released. This commit
overrides the original redline dependency with a jar released in the
folked corretto/redline repository. This allows us to unblock the build
on aarch64.