Integrate Cloudflare/Chromium Zlib
Integrate Cloudflare's optimized Zlib version from:
 https://github.com/cloudflare/zlib
Integrate Chromium's optimized Zlib version from:
 https://chromium.googlesource.com/chromium/src/third_party/zlib

Cloudflare's and Chromium's Zlib can be configured at build time with:
  --with-additional-zlib={cloudflare,chromium,none}

By default, Cloudflare/Chromium Zlib are enabled on all supported platforms.
But it is also possible to select only one of the additional libraries with:
  --with-additional-zlib={cloudflare|chromium}
or to exclude both versions on supported platforms with:
  --with-additional-zlib=none

At runtime, it is possible to choose any of the additionally configured
implementations with the new system properties:
   com.amazon.corretto.zlib.implementation
   com.amazon.corretto.zlib.implementation.inflate
   com.amazon.corretto.zlib.implementation.deflate
which can each be set to one of the following values:
   <cloudflare|chromium|bundled|system>

'bundled' chooses the original, bundled version of Zlib or the system
wide version if there is no bundled one (i.e the JDK was configured
with '--with-zlib=system').

'system' chooses the Zlib from your system library path if a version
can be found in your current library path. Otherwise, it falls back to
'bundled'.

The second and third property can be used to change the Inflate/Deflate
implementation selectively.

The default for all three properties is 'bundled'. The properties have
to be specified on the command line when launching java. They are only
evaluated at VM startup. Setting or changing the the properties later
will have no effect.